requirements.txt
django==3.1.5
asgiref==3.3.1
djangorestframework==3.8.2
pytz==2020.5
sqlparse==0.4.1
